requirements.txt
Akkoma.py==0.1.6
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
decorator==5.1.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
typing-extensions==4.12.2
urllib3==2.2.2
